src/dispatcher/requirements.txt
google-cloud-documentai>=2.24.0
functions-framework>=3.5.0
